+++ src/combinator/loop.ts
@@ -1,22 +1,32 @@
 ﻿import { Parser, Result } from './parser';
 
-export function loop<P extends Parser<any, any>[], R>(parser: Parser<R, P>, until?: string | RegExp): Parser<R, P> {
+export function loop<P extends Parser<any, any>[], R>(parser: Parser<R, P>, until?: string | RegExp | ((rest: string) => boolean)): Parser<R, P> {
+  const check = ((): (rest: string) => boolean => {
+    switch (typeof until) {
+      case 'undefined':
+        return _ => false;
+      case 'function':
+        return rest => (until as any)(rest);
+      default:
+        return rest => rest.slice(0, 99).search(until as string | RegExp) === 0;
+    }
+  })();
   return (source: string): Result<R, P> => {
     let rest = source;
-    if (until && rest.slice(0, 99).search(until) === 0) return;
     const results: R[] = [];
     while (true) {
       if (rest === '') break;
-      const r = parser(rest);
-      if (!r) break;
-      void results.push(...r[0]);
-      assert(r[1].length < rest.length);
-      assert(rest.endsWith(r[1]));
-      rest = r[1];
-      if (until && rest.slice(0, 99).search(until) === 0) break;
+      if (check(rest)) break;
+      const result = parser(rest);
+      if (!result) break;
+      const [rs, r] = result;
+      void results.push(...rs);
+      assert(r.length < rest.length);
+      assert(rest.endsWith(r));
+      rest = r;
     }
-    return rest.length < source.length
-        && (!until || rest.slice(0, 99).search(until) === 0)
+    if (rest === source) return;
+    return !until || check(rest)
       ? [results, rest]
       : void 0;
   };