+++ src/parser/inline/emphasis.ts
@@ -2,6 +2,7 @@
 import { combine } from '../../combinator/combine';
 import { loop } from '../../combinator/loop';
 import { EmphasisParser, InlineParser, inline } from '../inline';
+import { strong } from './strong';
 import { squash } from '../squash';
 
 type SubParsers = [InlineParser];
@@ -11,7 +12,7 @@ const closer = /^\*/;
 
 export const emphasis: EmphasisParser = function (source: string): Result<HTMLElement, SubParsers> {
   if (!source.startsWith('*') || source.search(syntax) !== 0) return;
-  const [cs, rest] = loop(combine<SubParsers, HTMLElement | Text>([inline]), closer)(source.slice(1)) || [[], ''];
+  const [cs, rest] = loop(combine<SubParsers, HTMLElement | Text>([inline]), rest => rest.slice(0, 99).search(closer) === 0 && !strong(rest))(source.slice(1)) || [[], ''];
   if (!rest.startsWith('*')) return;
   const el = document.createElement('em');
   void el.appendChild(squash(cs));